+++ src/smartcar/Sensor/SensorGyroData.java
@@ -13,14 +13,14 @@ import javax.swing.JButton;
  */
 public class SensorGyroData {
 
-    private float hori_angleSpeed;
-    private float hori_angle;
+    private double hori_angleSpeed;
+    private double hori_angle;
 
     public SensorGyroData() {
         this(0, 0);
     }
 
-    public SensorGyroData(float horiAngleSpeed, float horiAngle) {
+    public SensorGyroData(double horiAngleSpeed, double horiAngle) {
         this.hori_angleSpeed = horiAngleSpeed;
         this.hori_angle = horiAngle;
     }
@@ -28,14 +28,14 @@ public class SensorGyroData {
     /**
      * @return the hori_angleSpeed
      */
-    public float getHori_angleSpeed() {
+    public double getHori_angleSpeed() {
         return hori_angleSpeed;
     }
 
     /**
      * @param hori_angleSpeed the hori_angleSpeed to set
      */
-    public void setHori_angleSpeed(float hori_angleSpeed) {
+    public void setHori_angleSpeed(double hori_angleSpeed) {
         this.hori_angleSpeed = hori_angleSpeed;
     }
 
@@ -46,14 +46,14 @@ public class SensorGyroData {
     /**
      * @return the hori_angle
      */
-    public float getHori_angle() {
+    public double getHori_angle() {
         return hori_angle;
     }
 
     /**
      * @param hori_angle the hori_angle to set
      */
-    public void setHori_angle(float hori_angle) {
+    public void setHori_angle(double hori_angle) {
         this.hori_angle = hori_angle;
     }
 }
